Implement editing of scanning profiles
